+++ core/common/build.gradle.kts
@@ -9,8 +9,19 @@ apply<CommonGradlePlugin>()
 
 android {
     namespace = "team.retum.common"
+    buildFeatures {
+        compose = true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion = ProjectProperties.COMPOSE_COMPILER_EXTENSION
+    }
 }
 
 dependencies {
+
+    implementation(project(":core:design-system"))
+
     implementation(libs.androidx.lifecycle.viewmodel)
+    implementation(libs.androidx.compose.foundation)
+    implementation(libs.androidx.compose.material3)
 }
